refactor: rename abstract.repository.ts to repository.service.ts and update import paths
