+++ src/modules/projects/projects.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Project } from './entities/project.entity';
-import { RepositoryService } from 'src/core/repository/abstract.repository';
+import { RepositoryService } from '../../core/repository/repository.service';
 
 @Injectable()
 export class ProjectRepositoryService extends RepositoryService<Project> {
@@ -16,5 +16,4 @@ export class ProjectRepositoryService extends RepositoryService<Project> {
   protected get repository(): Repository<Project> {
     return this.projectRepo;
   }
-
 }
