+++ src/modules/users/user.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
-import { RepositoryService } from 'src/core/repository/abstract.repository';
+import { RepositoryService } from '../../core/repository/repository.service';
 
 @Injectable()
 export class UserRepositoryService extends RepositoryService<User> {
@@ -16,5 +16,4 @@ export class UserRepositoryService extends RepositoryService<User> {
   protected get repository(): Repository<User> {
     return this.userRepo;
   }
-
 }
