+++ src/modules/votes/votes.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Vote } from './entities/vote.entity';
-import { RepositoryService } from 'src/core/repository/abstract.repository';
+import { RepositoryService } from '../../core/repository/repository.service';
 
 @Injectable()
 export class VoteRepositoryService extends RepositoryService<Vote> {
